requirements.txt
alembic==1.7.6
aniso8601==9.0.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
boto3==1.21.27
botocore==1.24.27
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Deprecated==1.2.13
Flask==2.0.3
Flask-JWT-Extended==4.3.1
Flask-Limiter==2.2.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==1.0.0
limits==2.4.0
lxml==4.8.0
Mako==1.1.6
MarkupSafe==2.1.0
marshmallow==3.14.1
numpy==1.22.3
packaging==21.3
pandas==1.4.1
passlib==1.7.4
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
randomtimestamp==2.2
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
urllib3==1.26.8
webargs==8.1.0
Werkzeug==2.0.3
wrapt==1.13.3
